requirements.txt
oslo.config
python-keystoneclient
python-neutronclient
python-novaclient
python-saharaclient
python-troveclient
python-zaqarclient
